research_v2/requirements.txt
streamlit
python-docx
pandas
numpy
scikit-learn
scipy
sentence-transformers
keybert
tqdm
torch